testsuite: Fix struct size check [PR116155]
The size of "struct only_fam_2" is dependent on the alignment of the
flexible array member "b", and not on the type of the preceding
bit-fields.  For most targets the two are equal.  But on default_packed
targets like pru-unknown-elf, the alignment of int is not equal to the
size of int, so the test failed.

Patch was suggested by Qing Zhao.  Tested on pru-unknown-elf and
x86_64-pc-linux-gnu.

	PR testsuite/116155

gcc/testsuite/ChangeLog:

	* c-c++-common/fam-in-union-alone-in-struct-1.c: Adjust
	check to account for default_packed targets.

Signed-off-by: Dimitar Dimitrov <dimitar@dinux.eu>