testsuite: Fix up pr116037.c test [PR116245]
The test FAILs on big endian targets, because VV is a vector of unsigned __int128
and VC vector of unsigned char and so ((VC) vv)[0] is 0x01 on little endian
but 0xff on big endian and PDP endian.
As I believe it is intentional to test it as it is written on little endian,
the following patch just adds another case for big endian and for other
endians instead of figuring out what exactly to fetch it fetches the whole
unsigned __int128 and casts it to unsigned char.  Not that pdp11 has
__int128 support...

2024-08-06  Jakub Jelinek  <jakub@redhat.com>

	PR rtl-optimization/116037
	PR testsuite/116245
	* gcc.dg/torture/pr116037.c (foo): Fix up for big end middle endian.