vect: Small C++11-ification of vect_vect_recog_func_ptrs
This is a small C++11-ificiation for the use of vect_vect_recog_func_ptrs.
Changes the loop into a range based loop which then we can remove the variable
definition of NUM_PATTERNS. Also uses const reference instead of a pointer.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	* tree-vect-patterns.cc (NUM_PATTERNS): Delete.
	(vect_pattern_recog_1): Constify and change
	recog_func to a reference.
	(vect_pattern_recog): Use range-based loop over
	vect_vect_recog_func_ptrs.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>