ada: Avoid creating continuation messages without an intended parent
The messages modified in this patch do not have a clear intended
parent. This causes a lot of issues when grouping continuation
messages together with their parent. This can be confusing as it
is not obvious what was the parent message that caused this
problem or in worst case scenarios the message not being printed
alltogether.

These modified messages do not seem to be related to any concrete
error message and thus should be treated as independent messages.

gcc/ada/

	* sem_ch12.adb (Abandon_Instantiation): Remove continuation
	characters from the error message.
	* sem_ch13.adb (Check_False_Aspect_For_Derived_Type): Remove
	continuation characters from the error message.
	* sem_ch6.adb (Assert_False): Avoid creating a continuation
	message without a parent. If no primary message is created then
	the message is considered as primary.

gcc/testsuite/ChangeLog:

	* gnat.dg/interface6.adb: Adjust test.